requirements.txt
google-generativeai
pandas
openpyxl
pypdf
pyinstaller